The astral-go pin moves to the reflection codec's field-shape rule
astral-go 9b60ac4 refuses a struct field no Blueprint describes, and
moves its own types off plain Go fields. astrald follows the field types
that changed: Query.QueryString is String32, services.Update.Available is
Bool, and log.Entry.Level is Uint8. The call sites convert at the
boundary and their behaviour is unchanged.

The pin is the head of astral-go PR #81, which is unmerged. It moves to
the merge commit before this lands.

astrald's own types with plain fields move in the next commit. This
commit alone builds and passes go test ./..., but those types refuse to
encode under the pinned rule.

+++ cmd/astral-logview/entry_filter.go
@@ -21,7 +21,7 @@ func (e EntryFilter) Filter(entry *log.Entry) bool {
 		return false
 	}
 
-	if entry.Level > uint8(e.Level) {
+	if uint8(entry.Level) > uint8(e.Level) {
 		return false
 	}
